move product spy datastore from product usecases tests to an in memory data source implementation
