+++ backend/src/data/in-memory-product-store/index.ts
@@ -1,0 +1,1 @@
+export * from './in-memory-product-store'
