+++ backend/src/usecases/load-products-from-catalog/load-products-from-catalog.test.ts
@@ -1,31 +1,10 @@
-import { ProductStore } from '../../data/ports'
 import { LoadProductsFromCatalog } from './load-products-from-catalog'
-import { Product } from '../../domain/entities/product';
-
-class ProductStoreSpy implements ProductStore {
-  addCallsCount = 0
-  catalog: Array<Product> = []
-
-  async save (product: Product): Promise<void> {
-    this. addCallsCount++
-    this.catalog.push(product)
-  }
-
-  async loadById (productId: string) {
-    const product = this.catalog.find(p => p.getId() === productId)
-    return product
-  }
-
-  async loadAll (): Promise<Product[]> {
-    const productList = this.catalog
-    return productList
-  }
-}
+import { InMemoryProductStore } from '../../data/in-memory-product-store'
 
 describe('Product manager tests', () => {
   
   it('should not load Product on init', () => {
-    const inventoryStore = new ProductStoreSpy()
+    const inventoryStore = new InMemoryProductStore()
     new LoadProductsFromCatalog(inventoryStore)
 
     // inventoryManager.add(new Product('product1', 10))
